Rename lib directory to utils and improve handling when not logged in yet
