+++ pages/_app.js
@@ -1,7 +1,7 @@
 import React from "react";
 import App from "next/app";
 import "../public/tailwind.css";
-import AuthProvider from "./../lib/authProvider";
+import AuthProvider from "./../utils/authProvider";
 
 class MyApp extends App {
   render() {
