+++ pages/api/me.js
@@ -1,10 +1,10 @@
-import auth0 from '../../lib/auth0'
+import auth0 from "../../utils/auth0";
 
 export default async function me(req, res) {
   try {
-    await auth0.handleProfile(req, res)
+    await auth0.handleProfile(req, res);
   } catch (error) {
-    console.error(error)
-    res.status(error.status || 500).end(error.message)
+    console.error(error);
+    res.status(error.status || 500).end(error.message);
   }
 }
